refactor: 컨벤션 적용
