+++ src/main/java/roomescape/admin/controller/AdminController.java
@@ -22,7 +22,7 @@ public class AdminController {
 
     @GetMapping("/time")
     public String time(
-    ){
+    ) {
         return "/admin/time.html";
     }
 }
