+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.reservation.dao.ReservationDao;
 import roomescape.reservation.dto.ReservationRequest;
 import roomescape.reservation.dto.ReservationResponse;
-import roomescape.reservation.entity.Reservation;
 import roomescape.reservation.service.ReservationService;
 
 @RestController
