+++ src/main/java/roomescape/reservation/dao/ReservationDao.java
@@ -6,8 +6,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.reservation.entity.Reservation;
-import roomescape.reservation.entity.ReservationTime;
+import roomescape.reservation.model.Reservation;
+import roomescape.reservation.model.ReservationTime;
 
 @Repository
 public class ReservationDao {
@@ -28,7 +28,7 @@ public class ReservationDao {
         jdbcTemplate.update(connection -> {
             PreparedStatement ps = connection.prepareStatement(
                     sql,
-                    new String[]{"id"});
+                    new String[] {"id"});
             ps.setString(1, reservation.getName());
             ps.setString(2, reservation.getDate().toString());
             ps.setLong(3, reservation.getReservationTime().getId());
@@ -42,16 +42,16 @@ public class ReservationDao {
 
     public List<Reservation> findAllReservations() {
         String sql = """
-                SELECT
-                r.id as reservation_id,
-                r.name,
-                r.date,
-                t.id as time_id,
-                t.start_at as time_value
-            FROM reservation as r
-            inner join reservation_time as t
-            on r.time_id = t.id
-            """;
+                    SELECT
+                    r.id as reservation_id,
+                    r.name,
+                    r.date,
+                    t.id as time_id,
+                    t.start_at as time_value
+                FROM reservation as r
+                inner join reservation_time as t
+                on r.time_id = t.id
+                """;
 
         return jdbcTemplate.query(
                 sql,
@@ -61,8 +61,8 @@ public class ReservationDao {
                             resultSet.getString("name"),
                             resultSet.getDate("date").toLocalDate(),
                             new ReservationTime(
-                                resultSet.getLong("time_id"),
-                                resultSet.getTime("start_at").toLocalTime()
+                                    resultSet.getLong("time_id"),
+                                    resultSet.getTime("start_at").toLocalTime()
                             ));
                 });
     }
