+++ src/main/java/roomescape/reservation/model/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.entity;
+package roomescape.reservation.model;
 
 import java.time.LocalDate;
